requirements.txt
click==8.0.1
flake8==3.9.2
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
pymongo==3.12.0
python-dotenv==0.19.0
Werkzeug==2.0.1
